Move the structured extension codec out of a package cycle
The renderer lives in core.util and pulling the codec in from
core.certificate.request closed a cycle: that package already depends on
core.util in four files. The codec's nearest peer is the renderer, so core.util
is the better home and the edge disappears.

Also name the requested items when a structured extension no definition maps is
rejected, matching the RDN, SAN and extension messages, and document that an item
the platform cannot name fails closed even under extensibleList.

Refs #1950
