+++ src/main/java/com/otilm/core/attribute/CertificateRequestAttributeProjector.java
@@ -19,9 +19,9 @@ import com.otilm.api.model.core.certificate.CertificateType;
 import com.otilm.api.model.core.certificate.GeneralNameType;
 import com.otilm.api.model.core.oid.ExtensionValueEncoding;
 import com.otilm.api.model.core.oid.OidCategory;
-import com.otilm.core.certificate.request.StructuredExtensionCodec;
 import com.otilm.core.oid.OidHandler;
 import com.otilm.core.oid.OidRecord;
+import com.otilm.core.util.StructuredExtensionCodec;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.HashMap;
