+++ src/main/java/com/otilm/core/attribute/engine/AttributeEngine.java
@@ -53,7 +53,6 @@ import com.otilm.core.attribute.engine.records.ObjectAttributeContent;
 import com.otilm.core.attribute.engine.records.ObjectAttributeContentDetail;
 import com.otilm.core.attribute.engine.records.ObjectAttributeContentInfo;
 import com.otilm.core.attribute.engine.records.ObjectAttributeDefinitionContent;
-import com.otilm.core.certificate.request.StructuredExtensionCodec;
 import com.otilm.core.dao.entity.AttributeContent2Object;
 import com.otilm.core.dao.entity.AttributeContentItem;
 import com.otilm.core.dao.entity.AttributeDefinition;
@@ -73,6 +72,7 @@ import com.otilm.core.util.AuthHelper;
 import com.otilm.core.util.SearchHelper;
 import com.otilm.core.util.SecretEncodingVersion;
 import com.otilm.core.util.SecretsUtil;
+import com.otilm.core.util.StructuredExtensionCodec;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
 import java.time.DateTimeException;
